#209: Add comments and move code to a more logical location
