+++ datagateway_api/common/icat/helpers.py
@@ -27,7 +27,7 @@ from datagateway_api.common.icat.filters import (
     PythonICATLimitFilter,
     PythonICATWhereFilter,
 )
-from datagateway_api.common.icat.icat_client_pool import ExtendedLRUCache
+from datagateway_api.common.icat.lru_cache import ExtendedLRUCache
 from datagateway_api.common.icat.query import ICATQuery
 
 
@@ -79,7 +79,20 @@ def requires_session_id(method):
 @cached(cache=ExtendedLRUCache())
 def get_cached_client(session_id, client_pool):
     """
-    TODO - Add docstring
+    Get a client from cache using session ID as the cache parameter (client_pool will
+    always be given the same object, so won't impact on argument hashing)
+
+    An available client is fetched from the object pool, given a session ID, and kept
+    around in this cache until it becomes 'least recently used'. At this point, the
+    session ID is flushed and the client is returned to the pool. More details about
+    client handling can be found in the README
+
+    TODO - Add details of client handling in the README
+
+    :param session_id: The user's session ID
+    :type session_id: :class:`str`
+    :param client_pool: Client object pool used to fetch an unused client
+    :type client_pool: :class:`ObjectPool`
     """
 
     # Get a client from the pool
